Refactor my package
