+++ src/ru/nsu/ccfit/zuev/osu/ResourceManager.java
@@ -3,7 +3,6 @@ package ru.nsu.ccfit.zuev.osu;
 import static android.graphics.Bitmap.Config.ARGB_8888;
 
 import android.content.Context;
-import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.graphics.Typeface;
 import android.util.Log;
@@ -12,7 +11,7 @@ import com.dgsrz.bancho.security.SecurityUtils;
 
 import com.reco1l.framework.data.IniReader;
 import com.reco1l.legacy.data.SkinConverter;
-import com.reco1l.legacy.engine.BlankTextureRegion;
+import com.reco1l.legacy.graphics.texture.BlankTextureRegion;
 import org.andengine.engine.Engine;
 import org.andengine.opengl.font.Font;
 import org.andengine.opengl.font.FontFactory;
