+++ src/ru/nsu/ccfit/zuev/osu/game/GameScene.java
@@ -13,8 +13,8 @@ import com.edlplan.osu.support.timing.controlpoint.ControlPoints;
 import com.reco1l.api.ibancho.RoomAPI;
 import com.reco1l.framework.lang.Execution;
 import com.reco1l.framework.lang.execution.Async;
-import com.reco1l.legacy.engine.BlankTextureRegion;
-import com.reco1l.legacy.engine.VideoSprite;
+import com.reco1l.legacy.graphics.texture.BlankTextureRegion;
+import com.reco1l.legacy.graphics.sprite.VideoSprite;
 import com.reco1l.legacy.ui.entity.InGameLeaderboard;
 import com.reco1l.legacy.Multiplayer;
 import com.reco1l.legacy.ui.multiplayer.RoomScene;
